vlib: fix bug in dig_del_area() - update centroid topo       V2_delete_line protype change to use integer for 'line'       various minor fixes
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@49812 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ include/defs/vector.h
@@ -493,9 +493,9 @@ int V2_read_next_line_nat(struct Map_info *, struct line_pnts *,
 int V2_read_next_line_ogr(struct Map_info *, struct line_pnts *,
 			  struct line_cats *);
 int V1_delete_line_nat(struct Map_info *, off_t);
-int V2_delete_line_nat(struct Map_info *, off_t);
+int V2_delete_line_nat(struct Map_info *, int);
 int V1_delete_line_ogr(struct Map_info *, off_t);
-int V2_delete_line_ogr(struct Map_info *, off_t);
+int V2_delete_line_ogr(struct Map_info *, int);
 int V1_restore_line_nat(struct Map_info *, off_t);
 int V2_restore_line_nat(struct Map_info *, int, off_t);
 off_t V1_write_line_nat(struct Map_info *, int, const struct line_pnts *,
